rbz_ai/list_my_models.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    print("GEMINI_API_KEY not found in environment")
else:
    genai.configure(api_key=api_key)
    print("--- AVAILABLE MODELS ---")
    try:
        for m in genai.list_models():
            if 'generateContent' in m.supported_generation_methods:
                print(m.name)
        import sys
        sys.stdout.flush()
    except Exception as e:
        print(f"Error listing models: {e}")
