rbz_ai/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pypdf==3.17.4
pytesseract==0.3.10
Pillow==10.0.0